refactor: Tidying up
Did some renaming we suggested, additionally sorted some of the migrations, further testing needed
