--- service/grails-app/views/rulesetOwner/_rulesetOwner.gson
@@ -1,12 +1,0 @@
-import groovy.transform.*
-import org.olf.RulesetOwner
-import groovy.transform.Field
-import org.grails.orm.hibernate.cfg.GrailsHibernateUtil
-
-def should_expand = []
-
-@Field RulesetOwner rulesetOwner
-
-rulesetOwner = GrailsHibernateUtil.unwrapIfProxy(rulesetOwner) as RulesetOwner
-
-json g.render(rulesetOwner, [expand: should_expand])
