--- service/grails-app/views/rulesetTemplate/_rulesetTemplate.gson
@@ -1,12 +1,0 @@
-import groovy.transform.*
-import org.olf.RulesetTemplate
-import groovy.transform.Field
-
-def should_expand = [
-  'rulesetTemplateStatus',
-  'serialRuleset',
-]
-
-@Field RulesetTemplate rulesetTemplate
-
-json g.render(rulesetTemplate, [expand: should_expand])
